triton_meta is intended to be passed directly to triton. Previous we were also putting other metadata into triton_meta; but we should split out the other metadata into a separate dict to avoid possible conficts in the future.

This PR splits out triton_meta and inductor_meta so we have a place to put additional metadata that isn't intended to be passed to triton.

Tests - wait for CI

cc voznesenskym penguinwu EikanWang jgong5 Guobing-Chen XiaobingSuper zhuhaozhe blzheng wenzhe-nrv jiayisunx peterbell10 ipiszy yf225 chenyang78 kadeng muchulee8 aakhundov ColinPeppler

[ghstack-poisoned]

+++ c10/core/Allocator.cpp
@@ -58,7 +58,7 @@ void reportMemoryUsageToProfiler(
 }
 
 void reportOutOfMemoryToProfiler(
-    int64_t alloc_size,
+    size_t alloc_size,
     size_t total_allocated,
     size_t total_reserved,
     Device device) {
@@ -73,7 +73,7 @@ void reportOutOfMemoryToProfiler(
 MemoryReportingInfoBase::MemoryReportingInfoBase() = default;
 
 void MemoryReportingInfoBase::reportOutOfMemory(
-    int64_t /*alloc_size*/,
+    size_t /*alloc_size*/,
     size_t /*total_allocated*/,
     size_t /*total_reserved*/,
     Device /*device*/) {}
